Move make_document in milli and deprecate obkv_to_json
