+++ crates/milli/Cargo.toml
@@ -48,6 +48,7 @@ memmap2 = "0.9.9"
 obkv = "0.3.0"
 once_cell = "1.21.3"
 ordered-float = "5.1.0"
+permissive-json-pointer = { path = "../permissive-json-pointer" }
 rayon = "1.11.0"
 roaring = { version = "0.10.12", features = ["serde"] }
 rstar = { version = "0.12.2", features = ["serde"] }
