+++ crates/milli/src/lib.rs
@@ -41,6 +41,7 @@ pub mod snapshot_tests;
 pub mod constants;
 mod fieldids_weights_map;
 pub mod progress;
+mod value_paths_visitor;
 
 use std::collections::{BTreeMap, HashMap};
 use std::convert::{TryFrom, TryInto};
@@ -53,12 +54,15 @@ pub use filter_parser::{Condition, FilterCondition, IndexFilterCondition, Span, 
 use fxhash::{FxHasher32, FxHasher64};
 pub use grenad::CompressionType;
 pub use must_stop_processing::MustStopProcessing;
+use permissive_json_pointer::contained_in;
 pub use search::new::{
     execute_search, filtered_universe, DefaultSearchLogger, SearchContext, SearchLogger,
     VisualSearchLogger,
 };
+use serde::de::DeserializeSeed as _;
 use serde_json::Value;
 pub use thread_pool_no_abort::{CaughtPanic, ThreadPoolNoAbort, ThreadPoolNoAbortBuilder};
+use value_paths_visitor::ValuePathsVisitor;
 pub use {arroy, cellulite, charabia as tokenizer, hannoy, heed, rhai};
 
 pub use self::asc_desc::{AscDesc, AscDescError, Member, SortError};
@@ -253,6 +257,7 @@ pub fn bucketed_position(relative: u16) -> u16 {
     }
 }
 
+// TODO: deprecate this function in favor of `some_documents` in routes/indexes/documents.rs
 /// Transform a raw obkv store into a JSON Object.
 pub fn obkv_to_json(
     displayed_fields: &[FieldId],
@@ -272,6 +277,34 @@ pub fn obkv_to_json(
             Ok((name.to_owned(), value))
         })
         .collect()
+}
+
+pub fn make_document<S, I>(
+    obkv: &obkv::KvReaderU16,
+    field_ids_map: &FieldsIdsMap,
+    selectors: impl IntoIterator<IntoIter = I>,
+) -> Result<serde_json::Map<String, Value>>
+where
+    S: AsRef<str>,
+    I: Clone + Iterator<Item = S>,
+{
+    let selectors = selectors.into_iter();
+    let mut document = serde_json::Map::new();
+
+    for (key, value_bytes) in obkv {
+        let key = field_ids_map.name(key).expect("Missing field name");
+        if !selectors.clone().any(|selector| contained_in(selector.as_ref(), key)) {
+            // If the key is not part of the selection, skip this value
+            continue;
+        }
+
+        let visitor = ValuePathsVisitor::new_from_path(selectors.clone(), key);
+        let mut deserializer = serde_json::de::Deserializer::from_slice(value_bytes);
+        let value = visitor.deserialize(&mut deserializer).map_err(InternalError::SerdeJson)?;
+        document.insert(key.to_string(), value);
+    }
+
+    Ok(document)
 }
 
 /// Transform every field of a raw obkv store into a JSON Object.
